Use double values instead of BigDecimal to avoid scaling

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/search/ExtendedLuceneIndexExtractor.java
@@ -78,7 +78,7 @@ public class ExtendedLuceneIndexExtractor {
 				nextParam.getValueHigh(), nextParam.getValueHighDateOrdinal()));
 
 		theNewParams.myQuantityParams.forEach(nextParam ->
-			retVal.addQuantityIndexData(nextParam.getParamName(), nextParam.getUnits(), nextParam.getSystem(), nextParam.getValue()));
+			retVal.addQuantityIndexData(nextParam.getParamName(), nextParam.getUnits(), nextParam.getSystem(), nextParam.getValue().doubleValue()));
 
 
 		if (!theNewParams.myLinks.isEmpty()) {
